requirements.txt
Django
python-dotenv
pillow
psycopg2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
setuptools
requests
coverage
black
flake8
drf_yasg